more shuffling
